Refactor shell-command calls
Since we are calling nmcli, trimming, etc. quite a lot, the nm/cmd
function saves a lot of time and is easier to read. It could also
serve the same purpose as the return-nmcli-output but with an easier
and more general way of calling it, making it useful in other situations
as well.